colpali-backend/requirements.txt
torch
Pillow
datasets
tqdm
litserve
transformers 